requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
Levenshtein==0.27.1
loguru==0.7.3
mysql-connector-python==9.2.0
openai==1.66.3
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyodbc==5.2.0
RapidFuzz==3.12.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
win32-setctime==1.2.0
antlr4-python3-runtime==4.13.2